CommunityController V3 Added + Required Pythob Library to use this Code
Please note that there is a bug with this Library that causes the Left analog stick to not be centered completely which will  cause some in game functions to not work properly. For example You can't warp or navigate the maps in either Breath of the Wild or Super Mario Odyssey, and in certain games such as Fortnite(nintendo switch edition) Your character will meandor forward constantly.

There is also another bug that causes the serial ports to close randomly. I've "fixed" this by using an shell script that starts and closes the scripts every ~3minutes to keep the serial connection alive. It's messy but it works. If anyone would like to fix this, please feel free to submit a pull request. Of course if there are any questions please do not hesitate to ask